requirements.txt
streamlit
pandas
plotly
numpy
fpdf2
scikit-learn 